# services/control-plane-api/openapi/control_plane_openapi_v1.4.yaml
openapi: 3.0.3
info:
  title: AI Digital Office OS Control Plane
  version: 1.4.0
  description: >
    Governed control-plane API for AI DIGITAL OFFICE OS. Every endpoint is
    tenant-scoped via the JWT 'tenant_id' claim, versioned under /v1, and
    returns the standard error envelope on any non-2xx response (see
    ErrorResponse schema). v1.4 closes the integrity-review findings against
    v1.3: full project/agent-version/run/deployment/usage/tool lifecycle
    coverage, an explicit auth contract (issuer/audience/JWKS/scopes), and
    a required Idempotency-Key header on every state-changing endpoint.
    This specification is product-agnostic: no endpoint, schema, or example
    is specific to any named downstream project (see blueprint clause 75).
servers:
  - url: https://api.digitaloffice.example/v1
    description: Production (placeholder host — real endpoint is an
      environment-specific choice, intentionally not fabricated here)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        OAuth2 client-credentials token for service/agent callers, or
        OAuth2 authorization-code token for human users (blueprint clause 70).
        Required claims: 'iss' (issuer — this Office's or an approved
        external IdP), 'aud' (this control plane), 'exp'/'iat' (max lifetime
        15 minutes for service/agent callers, 8 hours for human sessions),
        'tenant_id', 'scope' (granted permission scopes), and 'actor_type'
        (SERVICE | HUMAN). Clock-skew tolerance ±60 seconds. JWKS rotates on
        a maximum 90-day cycle with a minimum 7-day overlap window. Session
        and refresh-token lifecycle is delegated entirely to the issuing IdP
        (blueprint clause 71) — this API only validates presented tokens.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
      description: >
        Required on every state-changing (POST/PATCH/DELETE) request.
        Replaying the same key within 24 hours returns the original result
        rather than reprocessing (blueprint clauses 49, 63, 68).
  schemas:
    ErrorResponse:
      type: object
      required: [error_code, message, correlation_id, retryable]
      properties:
        error_code:
          type: string
          enum: [VALIDATION_ERROR, AUTHORIZATION_ERROR, POLICY_ERROR, TOOL_ERROR,
                 MODEL_ERROR, PROVIDER_ERROR, NETWORK_ERROR, TIMEOUT,
                 DATABASE_ERROR, ARTIFACT_ERROR, DEPLOYMENT_ERROR, UNKNOWN_ERROR]
        message: { type: string }
        correlation_id: { type: string }
        retryable: { type: boolean }
...
    ApprovalRequest:
      type: object
      required: [action, risk_level]
      properties:
        task_id: { type: string }
        agent_id: { type: string }
        action:
          type: string
          description: "Includes AGENT_ACTIVATE for the Factory→Production transition (blueprint clause 45)."
        risk_level: { type: string, enum: [GREEN, YELLOW, RED] }
        reason: { type: string }
    ApprovalRecord:
      allOf:
        - $ref: '#/components/schemas/ApprovalRequest'
        - type: object
          properties:
            request_id: { type: string }
            tenant_id: { type: string, format: uuid }
            requester: { type: string }
            approver: { type: string }
            decision: { type: string, enum: [APPROVED, DENIED, EXPIRED] }
            expires_at: { type: string, format: date-time }
            execution_result: { type: object }
            created_at: { type: string, format: date-time }
            decided_at: { type: string, format: date-time }
    Workflow:
      type: object
      properties:
        workflow_id: { type: string }
        tenant_id: { type: string, format: uuid }
        project_id: { type: string }
        workflow_type: { type: string }
        definition_version: { type: string }
        current_state: { type: object }
        status: { type: string, enum: [RUNNING, PAUSED, COMPLETED, FAILED, CANCELLED, ESCALATED] }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
...
  responses:
...
    NotFound:
      description: Resource not found (or not visible to this tenant)
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
paths:
...
  /workflows/{workflow_id}:
    get:
      summary: Get durable workflow state (reconstructed from workflow_history on demand)
      parameters:
        - { name: workflow_id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Workflow state
          content: { application/json: { schema: { $ref: '#/components/schemas/Workflow' } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /workflows/{workflow_id}/pause:
    post:
      summary: Pause a running workflow
      parameters:
        - { name: workflow_id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Workflow paused }
  /workflows/{workflow_id}/resume:
    post:
      summary: Resume a paused workflow
      parameters:
        - { name: workflow_id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Workflow resumed }
...
  /workflows:
    post:
      summary: Start a new durable workflow
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, workflow_type, definition_version]
              properties:
                project_id: { type: string }
                workflow_type: { type: string }
                definition_version: { type: string }
      responses:
        '201': { description: Workflow started, content: { application/json: { schema: { $ref: '#/components/schemas/Workflow' } } } }
  /workflows/{workflow_id}/cancel:
    post:
      summary: Cancel a workflow (writes a workflow_history CANCELLED event)
      parameters: [ { name: workflow_id, in: path, required: true, schema: { type: string } } ]
      responses: { '200': { description: Workflow cancelled } }
  /workflows/{workflow_id}/retry:
    post:
      summary: Retry the current failed step of a workflow
      parameters: [ { name: workflow_id, in: path, required: true, schema: { type: string } } ]
      responses: { '200': { description: Retry initiated } }
  /workflows/{workflow_id}/escalate:
    post:
      summary: Escalate a workflow to human review (creates an approval_requests row)
      parameters: [ { name: workflow_id, in: path, required: true, schema: { type: string } } ]
      responses: { '200': { description: Escalation created, content: { application/json: { schema: { $ref: '#/components/schemas/ApprovalRecord' } } } } }
...